Fix HTTP TURL generation mechanism
The StoRM BE used to rely on an external mapper, provided by gridhttps
service, to create a TURL when HTTP or HTTPS was requested as a
transport protocol. This functionality was provided by the
storm-gridhttps-plugin, which also took care of adding permissions for
all accessed files to the storm gridhttps user and check the service
availablity by issuing HTTP head requests.

Now the code has been simplified to:

- provide a default implementation of the http plugin which does not set
  any permissions on files, thus reducing the impact on the underlying
  filesystem metadata, under the assumption that HTTP services will run
  as user storm (this also simplifies configuration)

- generate HTTP(s) TURLs in the backend without requiring any call to
  external services