Fix join in status get query that makes rf fail
A regression was introduced in the query used to fetch the status of Get
requests from the database.

The query was previously modified in the assumption that for all request
the ID for status_Get and request_Get would be identical. Actually, it
can happen that the ID for status_Get and request_Get drift to
inconsistent values, and it is exactly for this reason that the table
has an additional column for the foreign key reference.

This commit provides a tentative fix for the above issue.